Add UserVoter

+++ src/Controller/UserController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Form\UserFormType;
 use App\Repository\UserRepository;
+use App\Security\Voter\UserVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -17,7 +18,13 @@ final class UserController extends AbstractController
 {
     #[Route(name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
-    {
+    {   
+        if (!$this->isGranted(UserVoter::MANAGE)) {
+            $this->addFlash('error', 'user.access_denied_index');
+
+            return $this->redirectToRoute('home');
+        }
+
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -25,7 +32,13 @@ final class UserController extends AbstractController
 
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, UserPasswordHasherInterface $userPasswordHasherInterface): Response
-    {
+    {   
+        if (!$this->isGranted(UserVoter::MANAGE)) {
+            $this->addFlash('error', 'user.access_denied_new');
+
+            return $this->redirectToRoute('app_user_index');
+        }
+
         $user = new User();
         $isEdit = false;
         $form = $this->createForm(UserFormType::class, $user, ['is_edit' => $isEdit]);
@@ -51,22 +64,32 @@ final class UserController extends AbstractController
 
     #[Route('/{id}/show', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
-    {
+    {   
+        if (!$this->isGranted(UserVoter::MANAGE)) {
+            $this->addFlash('error', 'user.access_denied_show');
+
+            return $this->redirectToRoute('app_user_index');
+        }
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, User $user, EntityManagerInterface $entityManager, UserPasswordHasherInterface $userPasswordHasherInterface): Response
-    {   
+    public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
+    {       
+        if (!$this->isGranted(UserVoter::MANAGE)) {
+            $this->addFlash('error', 'user.access_denied_edit');
+
+            return $this->redirectToRoute('app_user_index');
+        }
+
         $isEdit = true;
         $form = $this->createForm(UserFormType::class, $user, ['is_edit' => $isEdit]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $user->setPassword($userPasswordHasherInterface->hashPassword($user, $user->getPassword()));
-
             $entityManager->flush();
 
             $this->addFlash("success", "user.edited");
@@ -83,9 +106,17 @@ final class UserController extends AbstractController
 
     #[Route('/{id}/delete', name: 'app_user_delete')]
     public function delete(User $user, EntityManagerInterface $entityManager): Response
-    {
+    {   
+        if (!$this->isGranted(UserVoter::MANAGE)) {
+            $this->addFlash('error', 'user.access_denied_delete');
+
+            return $this->redirectToRoute('app_user_index');
+        }
+
         $entityManager->remove($user);
         $entityManager->flush();
+
+        $this->addFlash("success", "user.deleted");
 
         return $this->redirectToRoute('app_user_index');
     }
